Fix build for PDCurses.
PDCurses no longer provides the low-level terminfo function putp, which
is okay because a build of PDCurses only knows about its particular
platform, not multiple terminfo entries. In addition, I renamed the
Makefile variables that can be set when calling make. For example:

make CURSES_CFLAGS=-I/usr/local/include/xcurses \
    CURSES_LIBDIRS= \
    CURSES_LIBS=/usr/local/libXCurses.dylib

I tested this with PDCurses 3.9 for X11 and SDL2 on macOS 10.14.4. The
SDL2 implementation was unable to set foreground colors, but it had the
same problem with its own demos.